sinks/util/service/net/unix: use sinks/util/unix/UnixEither and move impls there
